+++ tests/test_monitor.py
@@ -195,7 +195,7 @@ def test_enable_receiving_mock(monitor):
 
 def test_enable_receiving_error_mock(context, monitor, socket_path):
     with py.test.patch_libudev('udev_monitor_enable_receiving') as func:
-        with mock.patch('pyudev._monitor.get_libudev_errno') as get_errno:
+        with mock.patch('pyudev.monitor.get_libudev_errno') as get_errno:
             get_errno.return_value = errno.ENOENT
             func.return_value = 1
             with py.test.raises(EnvironmentError) as exc_info:
@@ -248,7 +248,7 @@ def test_receive_device_mock(monitor):
 
 def test_receive_device_error_mock(monitor):
     with py.test.patch_libudev('udev_monitor_receive_device') as func:
-        with mock.patch('pyudev._monitor.get_libudev_errno') as errorfunc:
+        with mock.patch('pyudev.monitor.get_libudev_errno') as errorfunc:
             func.return_value = None
             errorfunc.return_value = 0
             with py.test.raises(EnvironmentError) as exc_info: